created home page with routing
