+++ src/App.js
@@ -1,48 +1,19 @@
-import { useState } from "react";
+import { Route, RouterProvider, createBrowserRouter, createRoutesFromElements } from "react-router-dom";
 import "./App.css";
+import ErrorPage from './pages/ErrorPage';
+import HomePage from "./pages/HomePage";
 
 function App() {
-  const [originalUrl, setOriginalUrl] = useState("");
-  const [shortenedUrl, setShortenedUrl] = useState("");
 
-  const handleInputChange = (e) => {
-    setOriginalUrl(e.target.value);
-  }
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-    console.log("Submitting URL:", originalUrl)
-
-    setOriginalUrl('');
-  }
-
+  const router = createBrowserRouter(
+    createRoutesFromElements(
+      <Route path="/" errorElement={<ErrorPage />}>
+        <Route index element={<HomePage />} />
+      </Route>
+    )
+  )
   return (
-    <div className="App">
-      <h1>URL Shortener</h1>
-      <form onSubmit={handleSubmit}>
-        <label htmlFor="url">Enter your URL: </label>
-        <textarea
-          id="originalUrl"
-          value={originalUrl}
-          onChange={handleInputChange}
-          placeholder="Copy your URL here"
-          rows="4"
-          cols="50"
-        />
-        <br />
-        <button type="submit">Shorten URL</button>
-      </form>
-
-      {shortenedUrl && (
-        <div>
-          <h2>Shortened URL:</h2>
-          <a href={shortenedUrl} target="_blank" rel="noopener noreferrer">
-            {shortenedUrl}
-          </a>
-        </div>
-      )}
-
-    </div>
+    <RouterProvider router={router} />
   );
 }
